implementa testes automatizados com JUnit, Mockito e Testcontainers, ajusta configurações de segurança e remove temporariamente dependência do spring ai

+++ src/main/java/br/com/cinecrew/cinecrew/model/Movie.java
@@ -35,7 +35,7 @@ public class Movie extends BaseEntity {
     private String posterUrl;
 
     @Column(name = "release_year")
-    private Integer releaseYear;
+    private Short releaseYear;
 
     @Builder.Default
     @OneToMany(mappedBy = "movie")
